split netlinker funcs to netlink.go
